Auth Changes

+++ src/store/auth/store/userSlice.js
@@ -3,6 +3,7 @@ import { setInitialSettings } from '../../settingsSlice';
 import history from '../../../configurations/@history';
 import { AUTH } from '../../../configurations/config';
 import { signOut } from 'firebase/auth';
+import { userLogoutService } from '../../../services/default/defaultService';
 
 const initialState = {
   role: [], // guest
@@ -18,7 +19,8 @@ const initialState = {
     city: '',
     zipCode: '',
     about: '',
-    isPublic: false
+    isPublic: false,
+    user_id: null
   },
   user: null,
   loginStatus: false
@@ -37,13 +39,18 @@ const userSlice = createSlice({
 
 export const { setUser, updateUserData, userLoggedOut } = userSlice.actions;
 
-export const logoutUser = () => (dispatch) => {
-  return signOut(AUTH)
-    .then(() => {
-      localStorage.clear();
-      dispatch(userLoggedOut());
-      dispatch(setInitialSettings());
-      history.push('/login');
+export const logoutUser = () => async (dispatch, getState) => {
+  const state = getState();
+  const userData = state.auth.user.data;
+  return await signOut(AUTH)
+    .then(async () => {
+      const data = await userLogoutService({ id: userData.user_id, email: userData.email });
+      if (data.status) {
+        localStorage.clear();
+        await dispatch(userLoggedOut());
+        await dispatch(setInitialSettings());
+        history.push('/login');
+      }
     })
     .catch((err) => {
       console.log(err);
